pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chat_search"
version = "0.1.0"
description = "Chat search tool"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.68.0",
    "fire>=0.7.0",
    "tqdm>=4.67.1",
    "numpy>=1.24.0",
    "uvicorn>=0.27.0",
    "requests>=2.31.0",
    "openai>=1.82.0",
    "python-dotenv>=1.0.0",
    "telethon>=1.42.0",
    "lxml>=6.0.2",
    "html2text>=2025.4.15",
    "beautifulsoup4>=4.14.3",
    "aiosqlite>=0.20.0",
    "types-tqdm>=4.67.0.20250809",
]

[dependency-groups]
dev = [
    "black==25.1.0",
    "mypy==1.16.0",
    "flake8==7.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "isort>=6.1.0",
]

[tool.flake8]
exclude = ["*/__init__.py", ".venv"]
ignore = "E203,F403,E501,SIM115,PIE786,W503"
max-line-length = 120
